Tests/LibWeb/Text/input/URL/set-href-with-valid-hostname-but-invalid-port.html

<a id="id" href="http://example.net/path">
<script src="../include.js"></script>
<script>
    test(() => {
        let a = document.getElementById('id');
        println(a.href);
        a.host = "example.com:65536";
        println(a.href);
    })
</script>
